directors-cut/requirements.txt
# Director's Cut - Dual MCP Server for Video Editing
# Core dependencies

# Gradio with MCP support
gradio>=5.32.0
gradio[mcp]

# FastMCP for editing server
fastmcp>=2.0.0

# Video downloading
yt-dlp>=2024.1.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# AI/ML
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.3.0

# Speech recognition for subtitles
openai-whisper>=20231117
# Alternative: faster-whisper for better performance
# faster-whisper>=0.10.0

# Audio processing
librosa>=0.10.0
pydub>=0.25.1

# Image processing for intro generation
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
httpx>=0.25.0
